requirements.txt
absl-py==0.8.1
asgiref==3.2.3
astor==0.8.0
boto==2.49.0
boto3==1.10.44
botocore==1.13.44
certifi==2019.11.28
chardet==3.0.4
DAWG-Python==0.7.2
dj-database-url==0.5.0
Django==3.0.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.8
grpcio==1.16.1
gunicorn==20.0.4
h5py==2.9.0
idna==2.8
intents-classifier==0.4
jmespath==0.9.4
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nltk==3.4.5
numpy==1.17.4
opt-einsum==3.1.0
protobuf==3.11.2
psycopg2==2.8.4
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pyreadline==2.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
PyYAML==5.2
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.3.2
six==1.13.0
sklearn==0.0
smart-open==1.9.0
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0b1
tensorflow-estimator==2.0.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==5.0.1
wincertstore==0.2
wrapt==1.11.2
